+++ packages/nns/src/governance.canister.spec.ts
@@ -1,16 +1,5 @@
-import {
-  ActorSubclass,
-  AnonymousIdentity,
-  SubmitResponse,
-  polling,
-  type Agent,
-  type RequestId,
-} from "@dfinity/agent";
+import { ActorSubclass, AnonymousIdentity } from "@dfinity/agent";
 import { InvalidAccountIDError, LedgerCanister } from "@dfinity/ledger-icp";
-import {
-  ManageNeuronResponse as PbManageNeuronResponse,
-  NeuronId as PbNeuronId,
-} from "@dfinity/nns-proto";
 import { Principal } from "@dfinity/principal";
 import { InvalidPercentageError } from "@dfinity/utils";
 import { mock } from "jest-mock-extended";
@@ -26,7 +15,6 @@ import type {
 } from "../candid/governance";
 import { Topic, Vote } from "./enums/governance.enums";
 import {
-  FeatureNotSupportedError,
   GovernanceError,
   InsufficientAmountError,
   UnrecognizedTypeError,
@@ -46,29 +34,6 @@ const unexpectedGovernanceError: GovernanceErrorDetail = {
 };
 
 describe("GovernanceCanister", () => {
-  const requestId = new ArrayBuffer(20) as RequestId;
-  const agentCallSuccessfulResponse = {
-    requestId,
-    response: {
-      ok: true,
-      status: 13,
-      statusText: "good",
-    },
-  } as SubmitResponse;
-  const newSpawnNeuronId = new PbNeuronId();
-  newSpawnNeuronId.setId("1234");
-  const spawnResponse = new PbManageNeuronResponse.SpawnResponse();
-  spawnResponse.setCreatedNeuronId(newSpawnNeuronId);
-  const successfulResponse = new PbManageNeuronResponse();
-  successfulResponse.setSpawn(spawnResponse);
-  // We only care about the specifics in the Spawn Response.
-  // Therefore, we don't need to create a valid and specific response for each call.
-  const spyPollForResponse = jest
-    .spyOn(polling, "pollForResponse")
-    .mockImplementation(
-      jest.fn().mockResolvedValue(successfulResponse.serializeBinary()),
-    );
-
   const error: GovernanceErrorDetail = {
     error_message: "Some error",
     error_type: 1,
@@ -432,34 +397,6 @@ describe("GovernanceCanister", () => {
       expect(service.list_neurons).toBeCalled();
       expect(neurons.length).toBe(1);
     });
-
-    it("list Hardware Wallet neurons", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-      await governance.listNeurons({ certified: true });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
-    });
-
-    it("should not support query neurons with hardware wallet (only update calls)", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      const call = async () =>
-        await governance.listNeurons({ certified: false });
-
-      await expect(call).rejects.toThrow(new FeatureNotSupportedError());
-    });
   });
 
   describe("GovernanceCanister.listProposals", () => {
@@ -714,23 +651,6 @@ describe("GovernanceCanister", () => {
         additionalDissolveDelaySeconds: 100000,
       });
       expect(service.manage_neuron).toBeCalled();
-    });
-
-    it("increases dissolve delay of the neuron from a Hardware Wallet successfully", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      await governance.increaseDissolveDelay({
-        neuronId: BigInt(1),
-        additionalDissolveDelaySeconds: 100000,
-      });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
     });
 
     it("throw error when increaseDissolveDelay fails with error", async () => {
@@ -1004,20 +924,6 @@ describe("GovernanceCanister", () => {
       expect(service.manage_neuron).toBeCalled();
     });
 
-    it("successfully joins CF from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      await governance.joinCommunityFund(BigInt(10));
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
-    });
-
     it("throws error if response is error", async () => {
       const neuronId = BigInt(10);
       const serviceResponse: ManageNeuronResponse = {
@@ -1152,22 +1058,6 @@ describe("GovernanceCanister", () => {
       expect(service.manage_neuron).toBeCalled();
     });
 
-    it("successfully adds hotkey to neuron from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      const neuronId = BigInt(10);
-      const principal = Principal.fromText("kb4lg-bqaaa-aaaab-qabfq-cai");
-      await governance.addHotkey({ neuronId, principal });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
-    });
-
     it("throws error if response is error", async () => {
       const neuronId = BigInt(10);
       const serviceResponse: ManageNeuronResponse = {
@@ -1201,22 +1091,6 @@ describe("GovernanceCanister", () => {
       const principal = Principal.fromText("kb4lg-bqaaa-aaaab-qabfq-cai");
       await governance.removeHotkey({ neuronId, principal });
       expect(service.manage_neuron).toBeCalled();
-    });
-
-    it("successfully removes hotkey to neuron from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      const neuronId = BigInt(10);
-      const principal = Principal.fromText("kb4lg-bqaaa-aaaab-qabfq-cai");
-      await governance.removeHotkey({ neuronId, principal });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
     });
 
     it("throws error if response is error", async () => {
@@ -1363,23 +1237,6 @@ describe("GovernanceCanister", () => {
         percentageToMerge: 50,
       });
       expect(service.manage_neuron).toBeCalled();
-    });
-
-    it("successfully merges maturity neurons from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      await governance.mergeMaturity({
-        neuronId: BigInt(10),
-        percentageToMerge: 50,
-      });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
     });
 
     it("throws error if percentage not valid", async () => {
@@ -1538,33 +1395,6 @@ describe("GovernanceCanister", () => {
       expect(response).toBe(neuronId);
     });
 
-    it("successfully spawns new neuron from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      const requestId = new ArrayBuffer(20) as RequestId;
-      const response = {
-        requestId,
-        response: {
-          ok: true,
-          status: 13,
-          statusText: "good",
-        },
-      } as SubmitResponse;
-      agent.call.mockResolvedValue(response);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      const neuronId = await governance.spawnNeuron({
-        neuronId: BigInt(10),
-        percentageToSpawn: 50,
-      });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
-      expect(neuronId).toBe(BigInt(newSpawnNeuronId.getId()));
-    });
-
     it("throws error if percentage not valid", async () => {
       const service = mock<ActorSubclass<GovernanceService>>();
 
@@ -1615,22 +1445,6 @@ describe("GovernanceCanister", () => {
         neuronId,
       });
       expect(service.manage_neuron).toBeCalled();
-    });
-
-    it("successfully disburses neuron from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      await governance.disburse({
-        neuronId: BigInt(10),
-      });
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
     });
 
     it("throws error if response is error", async () => {
@@ -1726,20 +1540,6 @@ describe("GovernanceCanister", () => {
       expect(service.manage_neuron).toBeCalled();
     });
 
-    it("successfully starts dissolving process from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      await governance.startDissolving(BigInt(10));
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
-    });
-
     it("throws error if response is error", async () => {
       const neuronId = BigInt(10);
       const serviceResponse: ManageNeuronResponse = {
@@ -1770,20 +1570,6 @@ describe("GovernanceCanister", () => {
       });
       const response = await governance.stopDissolving(neuronId);
       expect(service.manage_neuron).toBeCalled();
-    });
-
-    it("successfully stops dissolving process from Hardware Wallet", async () => {
-      const agent = mock<Agent>();
-      agent.call.mockResolvedValue(agentCallSuccessfulResponse);
-
-      const governance = GovernanceCanister.create({
-        agent,
-        hardwareWallet: true,
-      });
-
-      await governance.stopDissolving(BigInt(10));
-      expect(agent.call).toBeCalled();
-      expect(spyPollForResponse).toBeCalled();
     });
 
     it("throws error if response is error", async () => {