+++ CHANGELOG.md
@@ -8,11 +8,14 @@
 - Remove hardware wallet specific code paths from `@dfinity/ledger-icp`.
 - Remove hardware wallet specific options from LedgerCanister.
 - Remove dependency on `@dfinity/nns-proto` from `@dfinity/ledger-icp`.
+- Remove hardware wallet specific code and `@dfinity/nns-proto` dependency from `@dfinity/nns`.
 
 # Breaking changes
 
 - Protocol buffers for hardware wallet transactions are no longer supported.
   Internet Computer Ledger app 2.4.9 or later is now required.
+- GovernanceCanister.listNeurons no longer throws an error when called with
+  `certified: false` for hardware wallet transactions.
 
 # 2024.03.25-1430Z